Added Mathew Purver segmentor (RndTopics).
Code refactoring: put the gibbs sampler part in another class since it
is a general process for both topic tracking and Purver's segmentors.

Note: still trying to run the gibbs sampler with Z variables. If I
assume the true values from the synthetic doc class and only sample for
rho it seems to be working.