Added support to test RndTopicsModel with real documents.
Currently I am still trying to figure out if working or not. If I don't
run the sampler run for many iterations it seems to get good
segmentations many times. But if I run it for a high number of
iterations it seems to always converge to a bad segmentation always (the
full document).